trend-review-app/requirements.txt
fastapi
python-multipart
uvicorn
google-cloud-firestore
pydantic
Jinja2